Extract `sprite_type` to own module
